perf(spec-parser): separate api validation logic (#11250)
* perf(spec-parser): separate api validation logic

* perf: remove unused code, add missing test file

* perf: remove unused import

---------
