+++ packages/spec-parser/src/validators/teamsAIValidator.ts
@@ -1,0 +1,44 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+"use strict";
+
+import { OpenAPIV3 } from "openapi-types";
+import { ParseOptions, APIValidationResult, ErrorType, ProjectType } from "../interfaces";
+import { Validator } from "./validator";
+
+export class TeamsAIValidator extends Validator {
+  constructor(spec: OpenAPIV3.Document, options: ParseOptions) {
+    super();
+    this.projectType = ProjectType.TeamsAi;
+    this.options = options;
+    this.spec = spec;
+  }
+
+  validateAPI(method: string, path: string): APIValidationResult {
+    const result: APIValidationResult = { isValid: true, reason: [] };
+    method = method.toLocaleLowerCase();
+
+    // validate method and path
+    const methodAndPathResult = this.validateMethodAndPath(method, path);
+    if (!methodAndPathResult.isValid) {
+      return methodAndPathResult;
+    }
+
+    const operationObject = (this.spec.paths[path] as any)[method] as OpenAPIV3.OperationObject;
+
+    // validate operationId
+    if (!this.options.allowMissingId && !operationObject.operationId) {
+      result.reason.push(ErrorType.MissingOperationId);
+    }
+
+    // validate server
+    const validateServerResult = this.validateServer(method, path);
+    result.reason.push(...validateServerResult.reason);
+
+    if (result.reason.length > 0) {
+      result.isValid = false;
+    }
+
+    return result;
+  }
+}
