+++ packages/spec-parser/test/utils.test.ts
@@ -65,2656 +65,6 @@ describe("utils", () => {
     });
   });
 
-  describe("isSupportedApi", () => {
-    it("should return true if method is GET, path is valid, and parameter is supported", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                  required: true,
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if have no operationId with allowMissingId is false", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                  required: true,
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: false,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.MissingOperationId]);
-    });
-
-    it("should return true if method is POST, path is valid, and no required parameters", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is POST, path is valid, parameter is supported and only one required param in parameters but contains auth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            api_key2: {
-              type: "apiKey",
-              name: "api_key2",
-              in: "header",
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  api_key2: [],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.AuthTypeIsNotSupported]);
-    });
-
-    it("should return true if allowBearerTokenAuth is true and contains bearer token auth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            bearer_token1: {
-              type: "http",
-              scheme: "bearer",
-            },
-            bearer_token2: {
-              type: "http",
-              scheme: "bearer",
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  bearer_token2: [],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowBearerTokenAuth: true,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return true if allowAPIKeyAuth is true and contains apiKey auth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            api_key2: {
-              type: "apiKey",
-              name: "api_key2",
-              in: "header",
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  api_key2: [],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: true,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if allowAPIKeyAuth is true but contains multiple apiKey auth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            api_key2: {
-              type: "apiKey",
-              name: "api_key2",
-              in: "header",
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  api_key2: [],
-                  api_key: [],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: true,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.MultipleAuthNotSupported]);
-    });
-
-    it("should return true if allowOauth2 is true and contains aad auth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            oauth: {
-              type: "oauth2",
-              flows: {
-                authorizationCode: {
-                  authorizationUrl: "https://example.com/api/oauth/dialog",
-                  tokenUrl: "https://example.com/api/oauth/token",
-                  refreshUrl: "https://example.com/api/outh/refresh",
-                  scopes: {
-                    "write:pets": "modify pets in your account",
-                    "read:pets": "read your pets",
-                  },
-                },
-              },
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  oauth: ["read:pets"],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: true,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if allowAPIKeyAuth is true, allowOauth2 is false, but contain oauth", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            oauth: {
-              type: "oauth2",
-              flows: {
-                authorizationCode: {
-                  authorizationUrl: "https://example.com/api/oauth/dialog",
-                  tokenUrl: "https://example.com/api/oauth/token",
-                  refreshUrl: "https://example.com/api/outh/refresh",
-                  scopes: {
-                    "write:pets": "modify pets in your account",
-                    "read:pets": "read your pets",
-                  },
-                },
-              },
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  oauth: ["read:pets"],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: true,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.AuthTypeIsNotSupported]);
-    });
-
-    it("should return false if allowAPIKeyAuth is true, allowOauth2 is true, but not auth code flow", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            oauth: {
-              type: "oauth2",
-              flows: {
-                implicit: {
-                  authorizationUrl: "https://example.com/api/oauth/dialog",
-                  scopes: {
-                    "write:pets": "modify pets in your account",
-                    "read:pets": "read your pets",
-                  },
-                },
-              },
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  oauth: ["read:pets"],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: true,
-        allowMultipleParameters: false,
-        allowOauth2: true,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.AuthTypeIsNotSupported]);
-    });
-
-    it("should return true if allowAPIKeyAuth is true, allowOauth2 is true, but not auth code flow for teams ai project", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        components: {
-          securitySchemes: {
-            api_key: {
-              type: "apiKey",
-              name: "api_key",
-              in: "header",
-            },
-            oauth: {
-              type: "oauth2",
-              flows: {
-                implicit: {
-                  authorizationUrl: "https://example.com/api/oauth/dialog",
-                  scopes: {
-                    "write:pets": "modify pets in your account",
-                    "read:pets": "read your pets",
-                  },
-                },
-              },
-            },
-          },
-        },
-        paths: {
-          "/users": {
-            post: {
-              security: [
-                {
-                  oauth: ["read:pets"],
-                },
-              ],
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: true,
-        allowMultipleParameters: false,
-        allowOauth2: true,
-        projectType: ProjectType.TeamsAi,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return true if method is POST, path is valid, parameter is supported and only one required param in parameters", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: false,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is POST, path is valid, parameter is supported and both postBody and parameters contains required param", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ExceededRequiredParamsLimit]);
-    });
-
-    it("should return true if method is POST, path is valid, parameter is supported and both postBody and parameters contains multiple required param for copilot", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should support multiple required parameters", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should not support multiple required parameters count larger than 5", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["id1", "id2", "id3", "id4", "id5", "id6"],
-                      properties: {
-                        id1: {
-                          type: "string",
-                        },
-                        id2: {
-                          type: "string",
-                        },
-                        id3: {
-                          type: "string",
-                        },
-                        id4: {
-                          type: "string",
-                        },
-                        id5: {
-                          type: "string",
-                        },
-                        id6: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ExceededRequiredParamsLimit]);
-    });
-
-    it("should support multiple required parameters count larger than 5 for teams ai project", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["id1", "id2", "id3", "id4", "id5", "id6"],
-                      properties: {
-                        id1: {
-                          type: "string",
-                        },
-                        id2: {
-                          type: "string",
-                        },
-                        id3: {
-                          type: "string",
-                        },
-                        id4: {
-                          type: "string",
-                        },
-                        id5: {
-                          type: "string",
-                        },
-                        id6: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.TeamsAi,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should not support multiple required parameters count larger than 5 for copilot", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["id1", "id2", "id3", "id4", "id5", "id6"],
-                      properties: {
-                        id1: {
-                          type: "string",
-                        },
-                        id2: {
-                          type: "string",
-                        },
-                        id3: {
-                          type: "string",
-                        },
-                        id4: {
-                          type: "string",
-                        },
-                        id5: {
-                          type: "string",
-                        },
-                        id6: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is POST, but requestBody contains unsupported parameter and required", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "array",
-                          items: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.PostBodyContainsRequiredUnsupportedSchema]);
-    });
-
-    it("should return true if method is POST, but requestBody contains unsupported parameter and required but has default value", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "array",
-                          default: ["item"],
-                          items: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: true,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is POST, parameters contain nested object, and request body is not json", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: {
-                    type: "object",
-                    required: ["name"],
-                    properties: {
-                      name: {
-                        type: "object",
-                        properties: {
-                          id: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "string",
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      expect(reason).to.have.members([
-        ErrorType.ParamsContainsNestedObject,
-        ErrorType.PostBodySchemaIsNotJson,
-      ]);
-      expect(reason.length).equals(2);
-    });
-
-    it("should return false if method is POST, but requestBody contain nested object", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "object",
-                          properties: {
-                            id: {
-                              type: "string",
-                            },
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.RequestBodyContainsNestedObject]);
-    });
-
-    it("should return true if method is POST, path is valid, parameter is supported and only one required param in postBody", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is GET, path is valid, parameter is supported, but response is empty", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                  required: true,
-                },
-              ],
-              responses: {
-                400: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ResponseJsonIsEmpty]);
-    });
-
-    it("should return false if method is not GET or POST", () => {
-      const method = "PUT";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.MethodNotAllowed]);
-    });
-
-    it("should return false if path is not valid", () => {
-      const method = "GET";
-      const path = "/invalid";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.UrlPathNotExist]);
-    });
-
-    it("should return false if parameter is not supported and required", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "object" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ParamsContainRequiredUnsupportedSchema]);
-    });
-
-    it("should return false due to ignore unsupported schema type with default value", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "object", default: { name: "test" } },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.NoParameter]);
-    });
-
-    it("should return false if parameter is in header and required", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "header",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ParamsContainRequiredUnsupportedSchema]);
-    });
-
-    it("should return true if parameter is in header and required for copilot", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "header",
-                  required: true,
-                  schema: { type: "string" },
-                },
-                {
-                  in: "query",
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if there is no parameters", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.NoParameter]);
-    });
-
-    it("should return true if there is no parameters for copilot", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if parameters is null", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.NoParameter]);
-    });
-
-    it("should return false if has parameters but no 20X response", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  schema: { type: "object" },
-                },
-              ],
-              responses: {
-                404: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "string",
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-
-      // NoParameter because object is not supported and there is no required parameters
-      expect(reason).to.have.members([ErrorType.NoParameter, ErrorType.ResponseJsonIsEmpty]);
-      expect(reason.length).equals(2);
-    });
-
-    it("should return false if method is POST, but request body contains media type other than application/json", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                  "application/xml": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [
-        ErrorType.PostBodyContainMultipleMediaTypes,
-        ErrorType.ExceededRequiredParamsLimit,
-      ]);
-    });
-
-    it("should return true if method is POST, and request body contains media type other than application/json for teams ai project", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                  "application/xml": {
-                    schema: {
-                      type: "object",
-                      required: ["name"],
-                      properties: {
-                        name: {
-                          type: "string",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.TeamsAi,
-        allowMethods: ["get", "post"],
-      };
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-
-    it("should return false if method is POST, and request body schema is not object", () => {
-      const method = "POST";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            post: {
-              requestBody: {
-                content: {
-                  "application/json": {
-                    schema: {
-                      type: "string",
-                    },
-                  },
-                },
-              },
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.Copilot,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.PostBodySchemaIsNotJson]);
-    });
-
-    it("should return false if method is GET, but response body contains media type other than application/json", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                    "application/xml": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.SME,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid, reason } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, false);
-      assert.deepEqual(reason, [ErrorType.ResponseContainMultipleMediaTypes]);
-    });
-
-    it("should return true if method is GET, and response body contains media type other than application/json for teams ai project", () => {
-      const method = "GET";
-      const path = "/users";
-      const spec = {
-        servers: [
-          {
-            url: "https://example.com",
-          },
-        ],
-        paths: {
-          "/users": {
-            get: {
-              parameters: [
-                {
-                  in: "query",
-                  required: true,
-                  schema: { type: "string" },
-                },
-              ],
-              responses: {
-                200: {
-                  content: {
-                    "application/json": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                    "application/xml": {
-                      schema: {
-                        type: "object",
-                        properties: {
-                          name: {
-                            type: "string",
-                          },
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      };
-      const options: ParseOptions = {
-        allowMissingId: true,
-        allowAPIKeyAuth: false,
-        allowMultipleParameters: false,
-        allowOauth2: false,
-        projectType: ProjectType.TeamsAi,
-        allowMethods: ["get", "post"],
-      };
-
-      const { isValid } = Utils.isSupportedApi(method, path, spec as any, options);
-      assert.strictEqual(isValid, true);
-    });
-  });
-
   describe("getUrlProtocol", () => {
     it("should return the protocol of a valid URL", () => {
       const url = "https://example.com/path/to/file";
@@ -2738,238 +88,6 @@ describe("utils", () => {
       const url = "ftp://v1";
       const protocol = Utils.getUrlProtocol(url);
       expect(protocol).to.equal("ftp:");
-    });
-  });
-
-  describe("checkRequiredParameters", () => {
-    it("should valid if there is only one required parameter", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "string" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-    });
-
-    it("should valid if there are multiple required parameters", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: true, schema: { type: "string" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 2);
-      assert.strictEqual(result.optionalNum, 0);
-    });
-
-    it("should not valid if any required parameter is in header or cookie and is required", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "string" } },
-        { in: "header", required: true, schema: { type: "string" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, false);
-    });
-
-    it("should valid if parameter in header or cookie is required but have default value", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "string" } },
-        { in: "header", required: true, schema: { type: "string", default: "value" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      // header param is ignored
-      assert.strictEqual(result.requiredNum, 1);
-      assert.strictEqual(result.optionalNum, 1);
-    });
-
-    it("should treat required param with default value as optional param", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string", default: "value" } },
-        { in: "path", required: false, schema: { type: "string" } },
-        { in: "query", required: true, schema: { type: "string" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 1);
-      assert.strictEqual(result.optionalNum, 2);
-    });
-
-    it("should ignore required query param with default value and array type", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "string" } },
-        { in: "query", required: true, schema: { type: "array", default: ["item"] } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 1);
-      assert.strictEqual(result.optionalNum, 1);
-    });
-
-    it("should ignore in header or cookie if is not required", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "string" } },
-        { in: "header", required: false, schema: { type: "string" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 1);
-      assert.strictEqual(result.optionalNum, 1);
-    });
-
-    it("should return false if any schema is array and required", () => {
-      const paramObject = [
-        { in: "query", required: true, schema: { type: "string" } },
-        { in: "path", required: true, schema: { type: "array" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, false);
-    });
-
-    it("should return false if any schema is object and required", () => {
-      const paramObject = [
-        { in: "query", required: false, schema: { type: "string" } },
-        { in: "path", required: true, schema: { type: "object" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, false);
-    });
-
-    it("should return valid if any schema is object but optional", () => {
-      const paramObject = [
-        { in: "query", required: false, schema: { type: "string" } },
-        { in: "path", required: false, schema: { type: "object" } },
-      ];
-      const result = Utils.checkParameters(paramObject as OpenAPIV3.ParameterObject[], false);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 0);
-      assert.strictEqual(result.optionalNum, 1);
-    });
-  });
-
-  describe("checkPostBodyRequiredParameters", () => {
-    it("should return 0 for an empty schema", () => {
-      const schema = {};
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.requiredNum, 0);
-      assert.strictEqual(result.optionalNum, 0);
-    });
-
-    it("should treat required schema with default value as optional param", () => {
-      const schema = {
-        type: "object",
-        required: ["name"],
-        properties: {
-          name: {
-            type: "string",
-            default: "value",
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.requiredNum, 0);
-      assert.strictEqual(result.optionalNum, 1);
-      assert.strictEqual(result.isValid, true);
-    });
-
-    it("should return 1 if the schema has a required string property", () => {
-      const schema = {
-        type: "object",
-        required: ["name"],
-        properties: {
-          name: {
-            type: "string",
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.requiredNum, 1);
-      assert.strictEqual(result.optionalNum, 0);
-      assert.strictEqual(result.isValid, true);
-    });
-
-    it("should return 0 if the schema has an optional string property", () => {
-      const schema = {
-        type: "object",
-        properties: {
-          name: {
-            type: "string",
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.requiredNum, 0);
-      assert.strictEqual(result.optionalNum, 1);
-      assert.strictEqual(result.isValid, true);
-    });
-
-    it("should return the correct count for a nested schema", () => {
-      const schema = {
-        type: "object",
-        required: ["name", "address"],
-        properties: {
-          name: {
-            type: "string",
-          },
-          address: {
-            type: "object",
-            required: ["street"],
-            properties: {
-              street: {
-                type: "string",
-              },
-              city: {
-                type: "string",
-              },
-            },
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.requiredNum, 2);
-      assert.strictEqual(result.optionalNum, 1);
-      assert.strictEqual(result.isValid, true);
-    });
-
-    it("should return not valid for an unsupported schema type", () => {
-      const schema = {
-        type: "object",
-        required: ["name"],
-        properties: {
-          name: {
-            type: "array",
-            items: {
-              type: "string",
-            },
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.isValid, false);
-    });
-
-    it("should return valid for an unsupported schema type but it is required with default value", () => {
-      const schema = {
-        type: "object",
-        required: ["name"],
-        properties: {
-          name: {
-            type: "array",
-            default: ["item"],
-            items: {
-              type: "string",
-            },
-          },
-        },
-      };
-      const result = Utils.checkPostBody(schema as any);
-      assert.strictEqual(result.isValid, true);
-      assert.strictEqual(result.requiredNum, 0);
-      assert.strictEqual(result.optionalNum, 0);
     });
   });
 
@@ -3337,43 +455,6 @@ describe("utils", () => {
       } as any;
       const { json, multipleMediaType } = Utils.getResponseJson(operationObject);
       expect(json).to.deep.equal({});
-      expect(multipleMediaType).to.be.true;
-    });
-
-    it("should return JSON response for status code 200 with multiple media type when it is teams ai project", () => {
-      const operationObject = {
-        responses: {
-          "200": {
-            content: {
-              "application/json": {
-                schema: {
-                  type: "object",
-                  properties: {
-                    message: { type: "string" },
-                  },
-                },
-              },
-              "application/xml": {
-                schema: {
-                  type: "object",
-                  properties: {
-                    message: { type: "string" },
-                  },
-                },
-              },
-            },
-          },
-        },
-      } as any;
-      const { json, multipleMediaType } = Utils.getResponseJson(operationObject, true);
-      expect(json).to.deep.equal({
-        schema: {
-          type: "object",
-          properties: {
-            message: { type: "string" },
-          },
-        },
-      });
       expect(multipleMediaType).to.be.true;
     });
